requirements.txt
# Core dependencies
fastapi==0.68.2
uvicorn==0.15.0
pydantic==1.10.7
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.2
python-json-logger==2.0.7
pyyaml==6.0.1
docker==6.1.3
requests==2.31.0

# Development dependencies
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
isort==5.12.0
mypy==1.5.1
